+++ packages/toolpad-app/pages/app-canvas/[[...path]].tsx
@@ -1,15 +1,10 @@
-import type { GetServerSideProps, NextPage } from 'next';
+import type { NextPage } from 'next';
 import * as React from 'react';
-import AppCanvas, { AppCanvasProps } from '../../src/canvas';
+import AppCanvas from '../../src/canvas';
+import loadComponents from '../../src/runtime/loadDomComponents';
 
-export const getServerSideProps: GetServerSideProps<AppCanvasProps> = async () => {
-  return {
-    props: {
-      basename: `/app-canvas`,
-    },
-  };
-};
-
-const App: NextPage<AppCanvasProps> = (props) => <AppCanvas {...props} />;
+const App: NextPage<{}> = (props) => (
+  <AppCanvas {...props} loadComponents={loadComponents} basename="/app-canvas" />
+);
 
 export default App;
