--- packages/toolpad-app/src/runtime/AppLoading.tsx
@@ -1,6 +1,0 @@
-import { LinearProgress } from '@mui/material';
-import * as React from 'react';
-
-export default function AppLoading() {
-  return <LinearProgress />;
-}
